+++ exonum/src/node/mod.rs
@@ -582,6 +582,8 @@ impl NodeHandler {
             .map(|conn| conn.address)
             .collect();
 
+        trace!("peers: {:?}", peers);
+
         for address in peers {
             self.send_to_addr(&address, message);
         }